requirements.txt
mysql-connector-python==8.2.0
qrcode[pil]==7.4.2
pycryptodome==3.19.0
Pillow==10.1.0